+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,8 +4,8 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.adapter.HandlerAdapter;
-import nextstep.mvc.adapter.HandlerAdapterRegistry;
+import nextstep.mvc.handleradapter.HandlerAdapter;
+import nextstep.mvc.handleradapter.HandlerAdapterRegistry;
 import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.handlermapping.HandlerMappingRegistry;
 import nextstep.mvc.view.ModelAndView;
