+++ mvc/src/main/java/nextstep/mvc/handleradapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.adapter;
+package nextstep.mvc.handleradapter;
 
 import nextstep.mvc.view.ModelAndView;
