+++ mvc/src/main/java/nextstep/mvc/handleradapter/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.adapter;
+package nextstep.mvc.handleradapter;
 
 import java.util.ArrayList;
 import java.util.List;
