+++ mvc/src/test/java/nextstep/mvc/DispatcherServletTest.java
@@ -7,8 +7,8 @@ import static org.mockito.Mockito.mock;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.adapter.AnnotationHandlerAdapter;
-import nextstep.mvc.adapter.ManualHandlerAdapter;
+import nextstep.mvc.handleradapter.AnnotationHandlerAdapter;
+import nextstep.mvc.handleradapter.ManualHandlerAdapter;
 import nextstep.mvc.handlermapping.AnnotationHandlerMapping;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
